Don't try to delete the old index if the table doesn't exist

+++ admin/class-update-database.php
@@ -36,18 +36,25 @@ class Update_Database {
 	public function edac_update_database() {
 
 		global $wpdb;
-		$table_name = $wpdb->prefix . 'accessibility_checker';
-		$query      = $wpdb->prepare( 'SHOW TABLES LIKE %s', $wpdb->esc_like( $table_name ) );
-		$db_version = get_option( 'edac_db_version' );
+		$table_name   = $wpdb->prefix . 'accessibility_checker';
+		$table_exists = $wpdb->get_var( // phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- Safe variable used for table name, caching not required for one time operation.
+			$wpdb->prepare( 'SHOW TABLES LIKE %s', $wpdb->esc_like( $table_name ) )
+		) === $table_name;
+		$db_version   = get_option( 'edac_db_version' );
 
 		// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared, WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- Prepare above, Safe variable used for table name, caching not required for one time operation.
-		if ( EDAC_DB_VERSION !== $db_version || $wpdb->get_var( $query ) !== $table_name ) {
+		if ( EDAC_DB_VERSION !== $db_version || ! $table_exists ) {
 
 			// If going from db version of below 1.0.4 then drop the UNIQUE index on `id` column, it
 			// is replaced by a PRIMARY KEY that has a PRIMARY index that constrains uniqueness.
-			if ( version_compare( $db_version, '1.0.4', '<' ) ) {
-				// phpcs:ignore WordPress.DB.PreparedSQL, WordPress.DB.DirectDatabaseQuery.SchemaChange -- Swapping from UNIQUE to PRIMARY indexing.
-				$wpdb->query( "ALTER TABLE $table_name DROP INDEX id" );
+			if ( version_compare( $db_version, '1.0.4', '<' ) && $table_exists ) {
+
+				// does the index exist?
+				$index_exists = $wpdb->get_var(	"SHOW INDEX FROM $table_name WHERE Key_name = 'id'"); // phpcs:ignore
+				// If the index exists, drop it.
+				if ( $index_exists ) {
+					$wpdb->query( "ALTER TABLE $table_name DROP INDEX id" ); // phpcs:ignore
+				}
 			}
 
 			$charset_collate = $wpdb->get_charset_collate();
